+++ Extension/src/common/messages/constants.ts
@@ -47,7 +47,7 @@ import { type FilterMetadata } from '../../background/api/filters/main';
 import { type GetAllowlistDomainsResponse } from '../../background/services/allowlist';
 import { type GetUserRulesEditorDataResponse, type GetUserRulesResponse } from '../../background/services/userrules';
 import { type GetCustomFilterInfoResult } from '../../background/api/filters/custom';
-import { type ManualExtensionUpdateData } from '../../background/services/extension-update/types';
+import { type ManualUpdateMetadata } from '../../background/services/extension-update/types';
 
 export const APP_MESSAGE_HANDLER_NAME = 'app';
 
@@ -82,7 +82,7 @@ export enum MessageType {
     LoadCustomFilterInfo = 'loadCustomFilterInfo',
     SubscribeToCustomFilter = 'subscribeToCustomFilter',
     RemoveAntiBannerFilter = 'removeAntiBannerFilter',
-    GetIsEngineStarted = 'getIsEngineStarted',
+    GetIsAppInitialized = 'getIsAppInitialized',
     GetTabInfoForPopup = 'getTabInfoForPopup',
     ChangeApplicationFilteringPaused = 'changeApplicationFilteringPaused',
     OpenRulesLimitsTab = 'openRulesLimitsTab',
@@ -186,8 +186,8 @@ export type RemoveListenerMessage = {
     };
 };
 
-export type GetIsEngineStartedMessage = {
-    type: MessageType.GetIsEngineStarted;
+export type GetIsAppInitializedMessage = {
+    type: MessageType.GetIsAppInitialized;
 };
 
 export type GetTabInfoForPopupMessage = {
@@ -250,7 +250,7 @@ export type CheckExtensionUpdateMessageMv3 = {
 export type UpdateExtensionMessageMv3 = {
     type: MessageType.UpdateExtensionMv3;
     data: {
-        from: ManualExtensionUpdateData['pageToOpenAfterReload'];
+        from: ManualUpdateMetadata['pageToOpenAfterReload'];
     };
 };
 
@@ -669,8 +669,8 @@ export type MessageMap = {
         message: AddFilteringSubscriptionMessage;
         response: void;
     };
-    [MessageType.GetIsEngineStarted]: {
-        message: GetIsEngineStartedMessage;
+    [MessageType.GetIsAppInitialized]: {
+        message: GetIsAppInitializedMessage;
         response: boolean;
     };
     [MessageType.GetTabInfoForPopup]: {