+++ Extension/src/common/user-scripts-api/user-scripts-api-mv3.ts
@@ -45,7 +45,7 @@ export const isUserScriptsApiSupported = () => {
  */
 export const shouldShowUserScriptsApiWarning = (): boolean => {
     if (isUserScriptsApiSupported()) {
-        logger.debug('[ext.user-scripts-api-mv3]: User Scripts API permission is already granted');
+        logger.trace('[ext.user-scripts-api-mv3]: User Scripts API permission is already granted');
         return false;
     }