+++ Extension/src/pages/popup/components/Popup/Popup.tsx
@@ -26,7 +26,6 @@ import { observer } from 'mobx-react';
 import { PopupLayout } from 'popup-layout';
 
 import { Icons } from '../ui/Icons';
-import { SplashScreen } from '../SplashScreen';
 import { popupStore } from '../../stores/PopupStore';
 import { messenger } from '../../../services/messenger';
 import { useAppearanceTheme } from '../../../common/hooks/useAppearanceTheme';
@@ -39,6 +38,7 @@ import {
 } from '../../../../common/messages';
 import { logger } from '../../../../common/logger';
 import { useObservePopupHeight } from '../../hooks/useObservePopupHeight';
+import { AnimatedLoader } from '../AnimatedLoader';
 
 import '../../styles/main.pcss';
 import './popup.pcss';
@@ -46,9 +46,9 @@ import './popup.pcss';
 export const Popup = observer(() => {
     const {
         appearanceTheme,
-        isLoading,
-        isEngineStarted,
-        checkIsEngineStarted,
+        isAppInitialized,
+        fetchIsAppInitialized,
+        setIsAppInitialized,
         getPopupData,
         updateBlockedStats,
         isAndroidBrowser,
@@ -59,10 +59,10 @@ export const Popup = observer(() => {
     // retrieve init data
     useEffect(() => {
         (async () => {
-            await checkIsEngineStarted();
+            await fetchIsAppInitialized();
             await getPopupData();
         })();
-    }, [checkIsEngineStarted, getPopupData]);
+    }, [fetchIsAppInitialized, getPopupData]);
 
     /**
      * We are adding "android" class to html element
@@ -135,6 +135,7 @@ export const Popup = observer(() => {
                     break;
                 }
                 case MessageType.AppInitialized: {
+                    setIsAppInitialized(true);
                     getPopupData();
                     break;
                 }
@@ -148,17 +149,15 @@ export const Popup = observer(() => {
         return () => {
             messenger.onMessage.removeListener(messageHandler);
         };
-    }, [updateBlockedStats, getPopupData]);
-
-    const PopupContent = isLoading
-        ? <SplashScreen isEngineStarted={isEngineStarted} />
-        : <PopupLayout />;
+    }, [updateBlockedStats, getPopupData, setIsAppInitialized]);
 
     return (
         <>
             <CommonIcons />
             <Icons />
-            {PopupContent}
+            <AnimatedLoader isLoading={!isAppInitialized}>
+                <PopupLayout />
+            </AnimatedLoader>
         </>
     );
 });