+++ Extension/src/pages/popup/stores/PopupStore/PopupStore-common.ts
@@ -38,7 +38,7 @@ import {
 import { translator } from '../../../../common/translators/translator';
 import { type PromoNotification } from '../../../../background/storages';
 import {
-    AppState,
+    type AppState,
     appStateActor,
     AppStateEvent,
 } from '../../state-machines/app-state-machine';
@@ -77,17 +77,10 @@ export abstract class PopupStoreCommon {
     isFilteringPossible = true;
 
     /**
-     * Flag that indicates whether the filtering engine is started.
-     *
-     * Needed for splash screen displaying.
-     *
-     * If not set, equals to `null` which means that the engine state is unknown
-     * and empty screen (splash screen with no logo) should be displayed.
-     * If set to `true`, the engine is started and the splash screen should not be displayed.
-     * If set to `false`, the engine is not started and the splash screen should be displayed.
+     * Flag that indicates whether the application is initialized.
      */
     @observable
-    isEngineStarted: boolean | null = null;
+    isAppInitialized: boolean = false;
 
     /**
      * Flag that indicates whether the filtering on a website is disabled
@@ -171,15 +164,22 @@ export abstract class PopupStoreCommon {
     }
 
     /**
-     * Checks whether the filtering engine is started.
+     * Fetches the application initialized state from the background.
+     */
+    fetchIsAppInitialized = async () => {
+        const res = await messenger.getIsAppInitialized();
+
+        this.setIsAppInitialized(res);
+    };
+
+    /**
+     * Sets the application initialized flag.
+     *
+     * @param value Whether the application is initialized.
      */
     @action
-    checkIsEngineStarted = async () => {
-        const res = await messenger.getIsEngineStarted();
-
-        runInAction(() => {
-            this.isEngineStarted = res;
-        });
+    setIsAppInitialized = (value: boolean) => {
+        this.isAppInitialized = value;
     };
 
     /**
@@ -194,16 +194,6 @@ export abstract class PopupStoreCommon {
             appStateActor.send({ type: AppStateEvent.Enable });
         }
     };
-
-    /**
-     * Flag that indicates whether the popup is loading
-     * and a splash screen should be displayed.
-     */
-    @computed
-    get isLoading(): boolean {
-        return this.appState === AppState.Loading
-            && !this.isEngineStarted;
-    }
 
     /**
      * Retrieves and sets up popup data including browser detection, tab info, and initial state.