+++ Extension/src/pages/services/messenger/messenger-common.ts
@@ -602,13 +602,13 @@ export abstract class MessengerCommon {
     };
 
     /**
-     * Sends a message to the background page to check if the engine is started.
+     * Sends a message to the background page to check if the application is initialized.
      *
      * @returns Promise that resolves to a boolean value:
-     * true if the engine is started, false otherwise.
+     * true if the application is initialized, false otherwise.
      */
-    getIsEngineStarted = async (): Promise<ExtractMessageResponse<MessageType.GetIsEngineStarted>> => {
-        return this.sendMessage(MessageType.GetIsEngineStarted);
+    getIsAppInitialized = async (): Promise<ExtractMessageResponse<MessageType.GetIsAppInitialized>> => {
+        return this.sendMessage(MessageType.GetIsAppInitialized);
     };
 
     /**